Add an easy way to manually translate strings (when automatic translation is not available)

+++ ui/locale/update-language.sh
@@ -2,6 +2,11 @@
 
 locale_dir="$PWD/${BASH_SOURCE%/*}"
 
+if [ "$1" == "-f" ]; then
+   force_transcribe="-f"
+   shift
+fi
+
 # language and country code, such as en_GB, de_DE or fr_FR
 lang_code=$1
 shift
@@ -19,7 +24,7 @@ shift
 # note the name should be pre-translated and is shown in the languages menu
 declare -a tables
 declare -a names
-while test $# -gt 0; do
+while [ $# -gt 0 ]; do
   tables+=("$1")
   shift
   names+=("$1")
@@ -46,20 +51,21 @@ do
   if [ ! -f "${transcribed_file}" ]; then
     # file doesn't exist yet, so create it from template
     msginit -l "${locale}" --no-translator -i "${template_file}" -o "${transcribed_file}"
-    if [ ! -z "${last_translation}" ]; then
-      # if we already have a language translation, resue that too
-      cp "${last_translation}" "${translation_file}"
-    fi
   else
     # update the file in case there are template changes
     msgmerge --lang="${lang_code}" -U --previous --backup=none "${transcribed_file}" "${template_file}"
+  fi
+  if [ ! -f "${translation_file}" ] && [ ! -z "${last_translation}" ]; then
+    # if we already have a language translation, resue that
+    cp "${last_translation}" "${translation_file}"
   fi
 
   # find any non-braille strings and translate and add them to the translation
   python3 deepl-translate.py -i "${transcribed_file}" -o "${translation_file}" -l "${deepl_code}"
 
   # use liblouis to transcribe back into original file
-  python3 liblouis-transcribe.py -i "${translation_file}" -o "${transcribed_file}" -t "${table_code}" -n "${names[$i]}"
+  python3 liblouis-transcribe.py -i "${translation_file}" -o "${transcribed_file}" -t "${table_code}" -n "${names[$i]}" $force_transcribe
+  last_translation="${translation_file}"
 
   # compile the .mo from the .po
   pybabel compile -f -D canute -d "${locale_dir}" -l "${locale}" -i "${transcribed_file}" -o "${compiled_file}"
